Use heuristic method to further improve test5
Originally, test5 suffers from very dense bitmap. It takes max 64
iterations in a loop.

In this heuristic method, if a bitset is found too dense, it's
inverted and turned into a sparse bitset for better processing.

Although the time spent on writing out[] is unchangable, this
heuristic method still gives a little better performance.